+++ src/main/scala/org/podval/tools/backend/nonjvm/NonJvmTestAdapter.scala
@@ -1,0 +1,8 @@
+package org.podval.tools.backend.nonjvm
+
+import sbt.testing.Framework
+
+trait NonJvmTestAdapter:
+  def loadFrameworks(frameworkNames: List[List[String]]): List[Option[Framework]]
+
+  def close(): Unit