+++ src/main/scala/org/podval/tools/backend/scalajs/ModuleInitializer.scala
@@ -1,0 +1,8 @@
+package org.podval.tools.backend.scalajs
+
+class ModuleInitializer(
+  val moduleId: String,
+  val className: String,
+  val mainMethodName: Option[String],
+  val mainMethodHasArgs: Boolean
+)