+++ src/main/scala/org/podval/tools/scalaplugin/jvm/JvmTask.scala
@@ -1,5 +1,0 @@
-package org.podval.tools.scalaplugin.jvm
-
-import org.podval.tools.scalaplugin.BackendTask
-
-trait JvmTask extends BackendTask