+++ src/main/scala/org/podval/tools/scalaplugin/scalajs/ScalaJSTask.scala
@@ -1,6 +1,0 @@
-package org.podval.tools.scalaplugin.scalajs
-
-import org.podval.tools.node.TaskWithNode
-import org.podval.tools.scalaplugin.nonjvm.NonJvmTask
-
-trait ScalaJSTask extends NonJvmTask[ScalaJSLinkTask] with TaskWithNode