+++ src/main/scala/org/podval/tools/scalaplugin/scalanative/ScalaNativeTask.scala
@@ -1,5 +1,0 @@
-package org.podval.tools.scalaplugin.scalanative
-
-import org.podval.tools.scalaplugin.nonjvm.NonJvmTask
-
-trait ScalaNativeTask extends NonJvmTask[ScalaNativeLinkTask]