+++ src/test/scala/org/podval/tools/test/scalajstutorial/ScalaJSTutorialTest.scala
@@ -1,7 +1,7 @@
 package org.podval.tools.test.scalajstutorial
 
-import org.podval.tools.build.ScalaBackend
-import org.podval.tools.build.scalajs.ScalaJSBackend
+import org.podval.tools.backend.ScalaBackend
+import org.podval.tools.backend.scalajs.ScalaJSBackend
 import org.podval.tools.test.testproject.{Feature, Fixture, ForClass, GroupingFunSpec, TestProject}
 
 class ScalaJSTutorialTest extends GroupingFunSpec: